FIX: 소수점 추가 안되는 이슈 수정

See merge request s10-blockchain-contract-sub2/S10P22D210!102

+++ backend/src/main/java/org/ssafy/d210/members/entity/Members.java
@@ -59,10 +59,10 @@ public class Members extends BaseTime {
     private GenderType gender;
 
     @Column(name = "height")
-    private Long height;
+    private double height;
 
     @Column(name = "weight")
-    private Long weight;
+    private double weight;
 
     @Column(name = "location", length = 500)
     private String location;
